ksrc: move to klplib package
